extract failing Deno.test
